requirements.txt
asgiref==3.9.1
django==5.2.4 
sqlparse==0.5.3
python-decouple==3.8
djangorestframework
markdown
django-filter
pillow
drf-yasg[validation]
djangorestframework-simplejwt

django-mail-templated

black
flake8

pytest-django
pytest